requirements.txt
pymupdf
pytesseract
pillow
google-generativeai
streamlit 
pdf2image  
textwrap
re
sklearn